+++ starsim_examples/__init__.py
@@ -1,1 +1,2 @@
 from .diseases import *
+from .networks import *
