+++ starsim_examples/networks/__init__.py
@@ -1,0 +1,3 @@
+from .embedding import *
+from .spatial import *
+from .theoretical import *
